unused_files/requirements_pars.txt
selenium-wire
undetected-chromedriver
pandas
openpyxl
psycopg2-binary
python-dotenv
